requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
httpx==0.25.2
Pillow==10.2.0
cloudinary==1.38.0
python-dotenv==1.0.0
resend==1.0.0
